moved simulation_settings.hpp to holohover_common
